Re-export 3x asset. Fix scaling factor.
Re-export 3x asset with scaling factor of 1.5 (or 3/2) from 2x asset via TexturePacker.

But 3x asset will be scaled down by 15% (divide by 1.15) to fit iPhone 6+ screen at the end.
Thus the actual scaling factor in effect is 1.5/1.15 = 1.30434782609 (~1.30).
Then from scaling factor used in CAWSpriteCoreLayer to scale down from retina screen is 0.5.
It increases to 0.5*1.3 = 0.65 or 15%. This means in final, we have a bigger scaling factor;
that is 0.5 - 0.15 = 0.35.